o Added support for new snapping during insert for test, autofrom,   objects, fromula, diagram, table objects
svn path=/trunk/koffice/; revision=478288

+++ kpresenter/KPresenterPageIface.h
@@ -102,7 +102,7 @@ k_dcop:
     DCOPRef insertLineV( int x,int y, int h, int w, bool rev );
     DCOPRef insertLineD1( int x,int y, int h, int w, bool rev );
     DCOPRef insertLineD2( int x,int y, int h, int w, bool rev );
-    DCOPRef insertTextObject( int x,int y, int h, int w );
+    DCOPRef insertTextObject( double x, double y, double h, double w );
     DCOPRef insertPicture( const QString & file, double x, double y, double h, double w );
 
     void deSelectAllObj();
